Test IRewardsController (#174)
* surface test coverage for RewardsController
* uint256 => uint

+++ contracts/interfaces/ICollateralizedBasketToken.sol
@@ -4,9 +4,9 @@ pragma solidity ^0.8.16;
 import "@openzeppelin/contracts/token/ERC20/extensions/IERC20Metadata.sol";
 
 interface ICollateralizedBasketToken is IERC20Metadata {
-    function mint(address account_, uint256 amount_) external;
+    function mint(address account_, uint amount_) external;
 
-    function burn(uint256 amount) external;
+    function burn(uint amount) external;
 
-    function burnFrom(address account_, uint256 amount_) external;
+    function burnFrom(address account_, uint amount_) external;
 }
