+++ contracts/libraries/RewardsDataTypes.sol
@@ -6,17 +6,17 @@ import "../interfaces/rewards/IEACAggregatorProxy.sol";
 library RewardsDataTypes {
     struct RewardsConfigInput {
         uint88 emissionPerSecond;
-        uint256 totalStaked;
+        uint totalStaked;
         uint32 distributionEnd;
-        address asset; // hypervisor
-        address reward; // CBT, USDC, Governance token
+        address asset;
+        address reward;
         IEACAggregatorProxy rewardOracle;
     }
 
     struct UserAssetBalance {
         address asset;
-        uint256 userStake;
-        uint256 totalStaked;
+        uint userStake;
+        uint totalStaked;
     }
 
     struct UserData {
